fix (clearCode) : indentation and import library
