+++ frontend/package.json
@@ -23,6 +23,7 @@
     "next": "13.4.19",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-router-dom": "^6.16.0",
     "typescript": "5.2.2",
     "yup": "^1.3.1"
   },
