+++ frontend/src/components/index.tsx
@@ -5,7 +5,7 @@ export { default as WCircleIcon } from "./atoms/CircleIcon/circleIcon";
 export { default as WLink } from "./atoms/Link/link";
 export { default as WTag} from "./atoms/Tag/tag"
 // export molecule components
-export { default as titleSearch} from './atoms/SearchComponents/titleSearch';
+export { default as WLogo} from './atoms/Logo/titleSearch';
 
 export { default as WVerificationCode } from './molecules/VerificationCode/verificationCode';
